2003-10-01  Lutz Mueller <lutz@users.sourceforge.net>
	* exif/exif-i18n.c: Patch to fix 64bit related crash by
	  Marcus Meissner (#784852).

+++ exif/exif-i18n.c
@@ -3,6 +3,7 @@
 
 #include <iconv.h>
 #include <string.h>
+#include <sys/types.h>
 
 #undef MIN
 #define MIN(a, b)  (((a) < (b)) ? (a) : (b))
@@ -11,9 +12,9 @@ const char *
 exif_i18n_convert_utf8_to_lat1 (const char *in)
 {
 	static iconv_t tr = 0;
-	unsigned int t = (in ? strlen (in) : 0);
+	size_t t = (in ? strlen (in) : 0);
 	static char buf[2048];
-	unsigned int buf_size = sizeof (buf);
+	size_t buf_size = sizeof (buf);
 	char *out = buf;
 
 	if (!in) return NULL;
